+++ netskope_ztre/CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.1.0
+## Added
+- Added support for adding multiple hosts in the 'Add Host to Private App' action.
+- Added a check to execute the revert action only when the host is available.
+- Added support for fetching domains of multiple applications in a single API request.
+## Changed
+- Made the 'Protocol' parameter optional for the 'Add host to Private App' action.
+## Fixed
+- Fixed a bug for the revert action.
+
 # 1.0.0
 ## Added
 - Initial release.